refactor(structure): consolidate views under Views/, add Services/ and Tasks/Models/
Restructure the source tree per #398. Every view concern now lives under
Views/{App,MenuBar,Timer,Tasks,Stats,Settings}/ with per-tab Components/ and
per-provider setup folders; non-view state holders and services move to Services/;
value types move to Tasks/Models/.

- Rename TimerView -> MenuBarPopoverView (file + type; 2 call sites).
- Split TimerSharedViews into CircularTimerView / ControlButton / SessionStatsView.
- Swap ObsidianProvider's `import SwiftUI` for `import Observation` so no concrete
  provider imports SwiftUI.

Pure relocation with no behavior change. The Xcode project uses file-system-
synchronized groups, so no project file edits are required.

Closes #398
Closes #388
