+++ app/Taskmato/Views/Tasks/TasksTabSections.swift
@@ -1,5 +1,5 @@
 //
-//  TasksTabViewTypes.swift
+//  TasksTabSections.swift
 //  Taskmato
 //
