+++ app/Taskmato/Views/Tasks/TasksTabSortMenu.swift
@@ -1,5 +1,5 @@
 //
-//  TasksTabViewSortMenu.swift
+//  TasksTabSortMenu.swift
 //  Taskmato
 //
