+++ app/Taskmato/Views/Timer/Components/CircularTimerView.swift
@@ -1,5 +1,5 @@
 //
-//  TimerSharedViews.swift
+//  CircularTimerView.swift
 //  Taskmato
 //
 
@@ -43,52 +43,5 @@ struct CircularTimerView: View {
       }
     }
     .frame(width: ringDiameter, height: ringDiameter)
-  }
-}
-
-/// A compact icon-only button used in the timer controls row.
-struct ControlButton: View {
-
-  /// The accessibility label and tooltip for this button.
-  let label: String
-  /// SF Symbol name for the button icon.
-  let icon: String
-  let action: () -> Void
-
-  var body: some View {
-    Button(action: action) {
-      Label(label, systemImage: icon)
-        .labelStyle(.iconOnly)
-        .frame(width: 32, height: 32)
-    }
-    .buttonStyle(.bordered)
-    .help(label)
-  }
-}
-
-/// A compact summary row showing today's focus session count and total focused time.
-struct SessionStatsView: View {
-
-  /// Number of completed focus sessions today.
-  let count: Int
-  /// Total minutes of completed focus time today.
-  let minutes: Int
-
-  var body: some View {
-    HStack {
-      Label(sessionLabel, systemImage: "timer")
-      Spacer()
-      Label(minuteLabel, systemImage: "clock")
-    }
-    .font(.caption)
-    .foregroundStyle(.secondary)
-  }
-
-  private var sessionLabel: String {
-    count == 1 ? "1 session today" : "\(count) sessions today"
-  }
-
-  private var minuteLabel: String {
-    "\(minutes) min focused"
   }
 }
